+++ src/assets/env.template.js
@@ -8,4 +8,5 @@
     window['env']['APPWRITE_ENDPOINT'] = '${APPWRITE_ENDPOINT}'
     window['env']['APPWRITE_WEBSITE_URL'] = '${APPWRITE_WEBSITE_URL}'
     window['env']['APPWRITE_RUNS_COLLECTION_ID'] = '${APPWRITE_RUNS_COLLECTION_ID}'
+    window['env']['NPOINT_DOC_ID'] = '${NPOINT_DOC_ID}'
 })(this)