+++ src/types/environment.d.ts
@@ -6,6 +6,7 @@ export interface Environment {
     appwriteEndpoint: string
     appwriteWebsiteUrl: string
     appwriteRunsCollectionId: string
+    npointDocId: string
 }
 
 export type EnvironmentType = 'testing' | 'development' | 'staging' | 'production'
@@ -13,13 +14,14 @@ export type EnvironmentType = 'testing' | 'development' | 'staging' | 'productio
 declare global {
     interface Window {
         env?: {
-            CLIMATE_ACTION_API_URL?: string
             ENVIRONMENT_TYPE?: EnvironmentType
+            CLIMATE_ACTION_API_URL?: string
             ORS_API_KEY?: string
             APPWRITE_PROJECT_ID?: string
             APPWRITE_ENDPOINT?: string
             APPWRITE_WEBSITE_URL?: string
             APPWRITE_RUNS_COLLECTION_ID?: string
+            NPOINT_DOC_ID?: string
         }
     }
 }